Added removing items from the shop
Not all of the shop actions are working now, but I have added the ability to remove item from the shop, whenever user bought something from it. Additionally, the amount of money are decreasing, when the user buys something, so it will be shown immediately on the label in the tkinter window.